requirements.txt
Flask==0.11.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Migrate==1.8.0
Flask-Assets==0.11
Flask-Mail==0.9.1
Flask-Security==1.7.5
flask-login==0.3.2
bcrypt==2.0.0
psycopg2==2.6.1
cssmin==0.2.0
jsmin==2.2.1
simple-salesforce==0.69
requests==2.10.0
